% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.r
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean LOCATION_NAME Column}
\usage{
eq_location_clean(eq_dataset_location)
}
\arguments{
\item{eq_dataset_location}{A dataframe of NOAA significant earthquakes data}
}
\value{
A dataframe with a clean LOCATION_NAME Column
}
\description{
This function cleans the LOCATION_NAME column by stripping out
the country name (including the colon) and converts names to title case.
}
\examples{
\dontrun{
 eq_loc_data <- eq_location_clean(eq_raw_data)
}

}
