% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_geoms.r
\name{geom_timeline}
\alias{geom_timeline}
\title{Geom Timeline}
\usage{
geom_timeline(mapping = NULL, data = NULL, na.rm = TRUE,
  position = "identity", stat = "identity", show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings.}

\item{data}{The data to be displayed in this layer.}

\item{na.rm}{If FALSE, the default, missing values are removed with a warning.
If TRUE, missing values are silently removed.}

\item{position}{Position adjustment, either as a string, or the result of a call to
a position adjustment function.}

\item{stat}{The statistical transformation to use on the data for this layer, as a string.}

\item{show.legend}{Logical. Should this layer be included in the legends?}

\item{inherit.aes}{If FALSE, overrides the default aesthetics.}

\item{...}{Other arguments passed on to layer.}
}
\value{
A ggplot2 layer
}
\description{
Plotting a time line of earthquakes. Each point represents an earthquake.
Optional aesthetics include color, size, and alpha (for transparency).
The x aesthetic is a date and an optional y aesthetic is a factor indicating some
stratification in which case multiple time lines will be plotted for each level
of the factor (e.g. country)
}
\examples{
\dontrun{eq_data <- readr::read_delim('signif.txt', delim = '\\t')}
\dontrun{
    eq_data \%>\%
        eq_clean_data("./data/signif.txt") \%>\%
        dplyr::filter(lubridate::year(DATE) > 2010 & COUNTRY \%in\% c('INDIA', 'USA')) \%>\%
        ggplot2::ggplot(ggplot2::aes(x = DATE, y = COUNTRY,
                                     colour = DEATHS, size = EQ_PRIMARY)) +
        geom_timeline(alpha = 0.5) +
        theme_timeline

}

}
