% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_geoms.r
\name{geom_timelinelabel}
\alias{geom_timelinelabel}
\title{Geom Timelinelabel}
\usage{
geom_timelinelabel(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = TRUE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings.}

\item{data}{The data to be displayed in this layer.}

\item{stat}{The statistical transformation to use on the data for this layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of a call to
a position adjustment function.}

\item{na.rm}{If FALSE, the default, missing values are removed with a warning.
If TRUE, missing values are silently removed.}

\item{show.legend}{Logical. Should this layer be included in the legends?}

\item{inherit.aes}{If FALSE, overrides the default aesthetics.}

\item{...}{Other arguments passed on to layer.}
}
\value{
A ggplot2 layer
}
\description{
Add annotations to the earthquake data.
This geom adds a vertical line to each data point with a text annotation
(e.g. the location of the earthquake) attached to each line.
We can subset to n_max number of earthquakes, where we take the n_max largest (by magnitude)
earthquakes. Aesthetics are x, which is the date of the earthquake and label which takes
the column name from which annotations will be obtained.
}
\examples{
\dontrun{eq_data <- readr::read_delim('signif.txt', delim = '\\t')}
\dontrun{
    eq_data \%>\%
        eq_clean_data("./data/signif.txt") \%>\%
        dplyr::filter(lubridate::year(DATE) > 2010 & COUNTRY \%in\% c('INDIA', 'USA')) \%>\%
        eq_location_clean() \%>\%
        ggplot2::ggplot(ggplot2::aes(x = DATE, y = COUNTRY,
                                     colour = DEATHS, size = EQ_PRIMARY)) +
        geom_timeline(alpha = 0.5) +
        geom_timelinelabel(ggplot2::aes(label = LOCATION_NAME, n_max = 3)) +
        theme_timeline

}

}
