% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_geoms.r
\docType{data}
\name{GeomTimeLine}
\alias{GeomTimeLine}
\title{GeomTimeLine}
\format{An object of class \code{GeomTimeLine} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
GeomTimeLine
}
\description{
GeomTimeLine class construction
}
\keyword{datasets}
