% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_geoms.r
\docType{data}
\name{GeomTimeLineLabel}
\alias{GeomTimeLineLabel}
\title{GeomTimeLineLabel}
\format{An object of class \code{GeomTimeLineLabel} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
GeomTimeLineLabel
}
\description{
GeomTimeLineLabel class constraction
}
\keyword{datasets}
