% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_geoms.r
\docType{data}
\name{theme_timeline}
\alias{theme_timeline}
\title{Theme Timeline}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 61.}
\usage{
theme_timeline
}
\description{
A new theme for geom_timeline
}
\keyword{datasets}
